refactor: 모든 Entity의 id(PK)의 type을 int -> Long으로 변경

+++ healthper/src/main/java/umc/healthper/domain/Comment.java
@@ -5,8 +5,6 @@ import lombok.Setter;
 import umc.healthper.global.BaseTimeEntity;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
@@ -16,7 +14,7 @@ public class Comment extends BaseTimeEntity {
     @Id
     @GeneratedValue
     @Column(name = "comment_id")
-    private Integer id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
